Using encoding/binary to encode and decode fixed64 and fixed32 values.

Here we show that unsafe is not much faster than a byte by byte copy, if we add a range check hint
https://github.com/awalterschulze/gorangecheck

Thank you @diamondo25 for reporting this issue
https://github.com/gogo/protobuf/issues/333