@awalterschulze 2017-10-07
```
$benchcmp ./test/mixbench/unsafe_unmarshaler.txt ./test/mixbench/unmarshaler.txt
benchmark                                                                   old ns/op     new ns/op     delta
BenchmarkNidOptNativeProtoUnmarshal-8                                       425           437           +2.82%
BenchmarkNinOptNativeProtoUnmarshal-8                                       612           648           +5.88%
BenchmarkNidRepNativeProtoUnmarshal-8                                       2884          2864          -0.69%
BenchmarkNinRepNativeProtoUnmarshal-8                                       2809          2820          +0.39%
BenchmarkNidRepPackedNativeProtoUnmarshal-8                                 1806          1801          -0.28%
BenchmarkNinRepPackedNativeProtoUnmarshal-8                                 1769          1813          +2.49%
BenchmarkNidOptStructProtoUnmarshal-8                                       1820          1839          +1.04%
BenchmarkNinOptStructProtoUnmarshal-8                                       1932          2018          +4.45%
BenchmarkNidRepStructProtoUnmarshal-8                                       5074          5149          +1.48%
BenchmarkNinRepStructProtoUnmarshal-8                                       4891          5018          +2.60%
BenchmarkNidEmbeddedStructProtoUnmarshal-8                                  1034          1065          +3.00%
BenchmarkNinEmbeddedStructProtoUnmarshal-8                                  1061          1079          +1.70%
BenchmarkNidNestedStructProtoUnmarshal-8                                    11579         11788         +1.80%
BenchmarkNinNestedStructProtoUnmarshal-8                                    10862         11113         +2.31%
BenchmarkNidOptCustomProtoUnmarshal-8                                       202           199           -1.49%
BenchmarkCustomDashProtoUnmarshal-8                                         228           228           +0.00%
BenchmarkNinOptCustomProtoUnmarshal-8                                       245           246           +0.41%
BenchmarkNidRepCustomProtoUnmarshal-8                                       810           801           -1.11%
BenchmarkNinRepCustomProtoUnmarshal-8                                       812           810           -0.25%
BenchmarkNinOptNativeUnionProtoUnmarshal-8                                  73.2          80.9          +10.52%
BenchmarkNinOptStructUnionProtoUnmarshal-8                                  194           201           +3.61%
BenchmarkNinEmbeddedStructUnionProtoUnmarshal-8                             459           460           +0.22%
BenchmarkNinNestedStructUnionProtoUnmarshal-8                               315           318           +0.95%
BenchmarkTreeProtoUnmarshal-8                                               307           308           +0.33%
BenchmarkOrBranchProtoUnmarshal-8                                           782           801           +2.43%
BenchmarkAndBranchProtoUnmarshal-8                                          776           786           +1.29%
BenchmarkLeafProtoUnmarshal-8                                               218           221           +1.38%
BenchmarkDeepTreeProtoUnmarshal-8                                           565           564           -0.18%
BenchmarkADeepBranchProtoUnmarshal-8                                        734           734           +0.00%
BenchmarkAndDeepBranchProtoUnmarshal-8                                      1263          1259          -0.32%
BenchmarkDeepLeafProtoUnmarshal-8                                           476           473           -0.63%
BenchmarkNilProtoUnmarshal-8                                                137           138           +0.73%
BenchmarkNidOptEnumProtoUnmarshal-8                                         141           143           +1.42%
BenchmarkNinOptEnumProtoUnmarshal-8                                         200           203           +1.50%
BenchmarkNidRepEnumProtoUnmarshal-8                                         449           461           +2.67%
BenchmarkNinRepEnumProtoUnmarshal-8                                         459           445           -3.05%
BenchmarkNinOptEnumDefaultProtoUnmarshal-8                                  203           199           -1.97%
BenchmarkAnotherNinOptEnumProtoUnmarshal-8                                  201           199           -1.00%
BenchmarkAnotherNinOptEnumDefaultProtoUnmarshal-8                           205           203           -0.98%
BenchmarkTimerProtoUnmarshal-8                                              223           224           +0.45%
BenchmarkMyExtendableProtoUnmarshal-8                                       654           653           -0.15%
BenchmarkOtherExtenableProtoUnmarshal-8                                     1353          1349          -0.30%
BenchmarkNestedDefinitionProtoUnmarshal-8                                   919           936           +1.85%
BenchmarkNestedDefinition_NestedMessageProtoUnmarshal-8                     427           421           -1.41%
BenchmarkNestedDefinition_NestedMessage_NestedNestedMsgProtoUnmarshal-8     216           224           +3.70%
BenchmarkNestedScopeProtoUnmarshal-8                                        893           918           +2.80%
BenchmarkNinOptNativeDefaultProtoUnmarshal-8                                671           663           -1.19%
BenchmarkCustomContainerProtoUnmarshal-8                                    351           360           +2.56%
BenchmarkCustomNameNidOptNativeProtoUnmarshal-8                             446           455           +2.02%
BenchmarkCustomNameNinOptNativeProtoUnmarshal-8                             652           667           +2.30%
BenchmarkCustomNameNinRepNativeProtoUnmarshal-8                             2841          2908          +2.36%
BenchmarkCustomNameNinStructProtoUnmarshal-8                                2639          2668          +1.10%
BenchmarkCustomNameCustomTypeProtoUnmarshal-8                               919           926           +0.76%
BenchmarkCustomNameNinEmbeddedStructUnionProtoUnmarshal-8                   461           471           +2.17%
BenchmarkCustomNameEnumProtoUnmarshal-8                                     270           267           -1.11%
BenchmarkNoExtensionsMapProtoUnmarshal-8                                    353           341           -3.40%
BenchmarkUnrecognizedProtoUnmarshal-8                                       85.7          85.0          -0.82%
BenchmarkUnrecognizedWithInnerProtoUnmarshal-8                              415           408           -1.69%
BenchmarkUnrecognizedWithInner_InnerProtoUnmarshal-8                        42.6          45.5          +6.81%
BenchmarkUnrecognizedWithEmbedProtoUnmarshal-8                              259           261           +0.77%
BenchmarkUnrecognizedWithEmbed_EmbeddedProtoUnmarshal-8                     43.2          43.2          +0.00%
BenchmarkNodeProtoUnmarshal-8                                               306           310           +1.31%
BenchmarkNonByteCustomTypeProtoUnmarshal-8                                  308           310           +0.65%
BenchmarkNidOptNonByteCustomTypeProtoUnmarshal-8                            292           294           +0.68%
BenchmarkNinOptNonByteCustomTypeProtoUnmarshal-8                            306           308           +0.65%
BenchmarkNidRepNonByteCustomTypeProtoUnmarshal-8                            1049          1040          -0.86%
BenchmarkNinRepNonByteCustomTypeProtoUnmarshal-8                            1046          1039          -0.67%
BenchmarkProtoTypeProtoUnmarshal-8                                          221           220           -0.45%

benchmark                                                                   old MB/s     new MB/s     speedup
BenchmarkNidOptNativeProtoUnmarshal-8                                       540.14       525.94       0.97x
BenchmarkNinOptNativeProtoUnmarshal-8                                       341.01       322.37       0.95x
BenchmarkNidRepNativeProtoUnmarshal-8                                       283.98       285.90       1.01x
BenchmarkNinRepNativeProtoUnmarshal-8                                       291.49       290.41       1.00x
BenchmarkNidRepPackedNativeProtoUnmarshal-8                                 206.53       207.09       1.00x
BenchmarkNinRepPackedNativeProtoUnmarshal-8                                 210.78       205.70       0.98x
BenchmarkNidOptStructProtoUnmarshal-8                                       464.71       459.88       0.99x
BenchmarkNinOptStructProtoUnmarshal-8                                       396.30       379.50       0.96x
BenchmarkNidRepStructProtoUnmarshal-8                                       349.42       344.29       0.99x
BenchmarkNinRepStructProtoUnmarshal-8                                       362.49       353.26       0.97x
BenchmarkNidEmbeddedStructProtoUnmarshal-8                                  466.75       453.50       0.97x
BenchmarkNinEmbeddedStructProtoUnmarshal-8                                  431.37       424.46       0.98x
BenchmarkNidNestedStructProtoUnmarshal-8                                    356.66       350.34       0.98x
BenchmarkNinNestedStructProtoUnmarshal-8                                    358.12       350.03       0.98x
BenchmarkNidOptCustomProtoUnmarshal-8                                       350.76       355.77       1.01x
BenchmarkCustomDashProtoUnmarshal-8                                         358.70       359.40       1.00x
BenchmarkNinOptCustomProtoUnmarshal-8                                       272.41       271.79       1.00x
BenchmarkNidRepCustomProtoUnmarshal-8                                       224.66       227.02       1.01x
BenchmarkNinRepCustomProtoUnmarshal-8                                       224.11       224.64       1.00x
BenchmarkNinOptNativeUnionProtoUnmarshal-8                                  218.58       197.77       0.90x
BenchmarkNinOptStructUnionProtoUnmarshal-8                                  323.57       311.92       0.96x
BenchmarkNinEmbeddedStructUnionProtoUnmarshal-8                             324.42       323.74       1.00x
BenchmarkNinNestedStructUnionProtoUnmarshal-8                               247.13       245.16       0.99x
BenchmarkTreeProtoUnmarshal-8                                               335.37       333.66       0.99x
BenchmarkOrBranchProtoUnmarshal-8                                           313.01       305.77       0.98x
BenchmarkAndBranchProtoUnmarshal-8                                          315.38       311.55       0.99x
BenchmarkLeafProtoUnmarshal-8                                               443.71       437.02       0.98x
BenchmarkDeepTreeProtoUnmarshal-8                                           256.61       256.77       1.00x
BenchmarkADeepBranchProtoUnmarshal-8                                        249.27       249.07       1.00x
BenchmarkAndDeepBranchProtoUnmarshal-8                                      262.80       263.56       1.00x
BenchmarkDeepLeafProtoUnmarshal-8                                           294.03       295.86       1.01x
BenchmarkNilProtoUnmarshal-8                                                254.01       252.17       0.99x
BenchmarkNidOptEnumProtoUnmarshal-8                                         260.58       258.45       0.99x
BenchmarkNinOptEnumProtoUnmarshal-8                                         204.30       201.75       0.99x
BenchmarkNidRepEnumProtoUnmarshal-8                                         131.35       127.85       0.97x
BenchmarkNinRepEnumProtoUnmarshal-8                                         128.30       132.38       1.03x
BenchmarkNinOptEnumDefaultProtoUnmarshal-8                                  201.59       205.22       1.02x
BenchmarkAnotherNinOptEnumProtoUnmarshal-8                                  203.75       205.47       1.01x
BenchmarkAnotherNinOptEnumDefaultProtoUnmarshal-8                           199.16       201.78       1.01x
BenchmarkTimerProtoUnmarshal-8                                              469.62       468.19       1.00x
BenchmarkMyExtendableProtoUnmarshal-8                                       123.72       123.97       1.00x
BenchmarkOtherExtenableProtoUnmarshal-8                                     116.72       117.08       1.00x
BenchmarkNestedDefinitionProtoUnmarshal-8                                   252.43       247.70       0.98x
BenchmarkNestedDefinition_NestedMessageProtoUnmarshal-8                     278.50       282.31       1.01x
BenchmarkNestedDefinition_NestedMessage_NestedNestedMsgProtoUnmarshal-8     378.01       364.55       0.96x
BenchmarkNestedScopeProtoUnmarshal-8                                        249.72       242.79       0.97x
BenchmarkNinOptNativeDefaultProtoUnmarshal-8                                311.36       315.02       1.01x
BenchmarkCustomContainerProtoUnmarshal-8                                    310.13       302.50       0.98x
BenchmarkCustomNameNidOptNativeProtoUnmarshal-8                             514.92       504.73       0.98x
BenchmarkCustomNameNinOptNativeProtoUnmarshal-8                             320.55       313.20       0.98x
BenchmarkCustomNameNinRepNativeProtoUnmarshal-8                             288.23       281.59       0.98x
BenchmarkCustomNameNinStructProtoUnmarshal-8                                363.30       359.44       0.99x
BenchmarkCustomNameCustomTypeProtoUnmarshal-8                               232.73       231.07       0.99x
BenchmarkCustomNameNinEmbeddedStructUnionProtoUnmarshal-8                   322.60       316.30       0.98x
BenchmarkCustomNameEnumProtoUnmarshal-8                                     166.20       168.48       1.01x
BenchmarkNoExtensionsMapProtoUnmarshal-8                                    228.99       237.47       1.04x
BenchmarkUnrecognizedProtoUnmarshal-8                                       525.21       529.53       1.01x
BenchmarkUnrecognizedWithInnerProtoUnmarshal-8                              226.40       230.38       1.02x
BenchmarkUnrecognizedWithInner_InnerProtoUnmarshal-8                        117.47       109.87       0.94x
BenchmarkUnrecognizedWithEmbedProtoUnmarshal-8                              343.44       340.78       0.99x
BenchmarkUnrecognizedWithEmbed_EmbeddedProtoUnmarshal-8                     115.84       115.70       1.00x
BenchmarkNodeProtoUnmarshal-8                                               329.03       325.74       0.99x
BenchmarkNonByteCustomTypeProtoUnmarshal-8                                  256.11       254.37       0.99x
BenchmarkNidOptNonByteCustomTypeProtoUnmarshal-8                            284.10       281.53       0.99x
BenchmarkNinOptNonByteCustomTypeProtoUnmarshal-8                            258.01       256.49       0.99x
BenchmarkNidRepNonByteCustomTypeProtoUnmarshal-8                            219.19       221.03       1.01x
BenchmarkNinRepNonByteCustomTypeProtoUnmarshal-8                            219.68       221.20       1.01x
BenchmarkProtoTypeProtoUnmarshal-8                                          370.99       372.18       1.00x
```